# docs/reviews/project-category-only-evidence/current-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/links/order": {
      "put": {
        "tags": [
          "Links"
        ],
        "summary": "Reorder all owned Links",
        "description": "Clear filters and submit the exact whole-workspace ID permutation with collectionRevision. Stale revision 409 precedes ID-set validation 400. Assign dense positions; moved items advance revision. Even unchanged/empty valid order advances collectionRevision once.",
        "operationId": "reorder",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderLinksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Full unfiltered ordered collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkListResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ACCOUNT_DISABLED; authenticated mutations may also return CSRF_INVALID for a missing/invalid token or forbidden Origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND: missing Workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "REVISION_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/v1/tasks/{id}/restore": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Restore an owned Task",
        "description": "Body contains only revision. No creation Idempotency-Key is required; archived Project relations are retained.",
        "operationId": "restore",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Full restored Task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ACCOUNT_DISABLED; authenticated mutations may also return CSRF_INVALID for a missing/invalid token or forbidden Origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND: missing or inaccessible Task/Project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "REVISION_CONFLICT or INVALID_RESOURCE_STATE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/v1/links": {
      "get": {
        "tags": [
          "Links"
        ],
        "summary": "List owned Links",
        "description": "Full filtered collection, position ASC then id ASC; nextCursor always null. Scope unity/server includes common all links. Query is literal case-insensitive label/description/url search. Maximum 500 items / 8 MiB compact UTF-8; no limit/cursor/project parameters.",
        "operationId": "list_4",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all",
              "enum": [
                "all",
                "unity",
                "server"
              ]
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full filtered collection and same-snapshot global collectionRevision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkListResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ACCOUNT_DISABLED; authenticated mutations may also return CSRF_INVALID for a missing/invalid token or forbidden Origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "Links"
        ],
        "summary": "Create an owned Link",
        "description": "Append to global order. Required same key/body replays original wrapper for 24 hours even after deletion. Maximum configured 500 live Links.",
        "operationId": "create_4",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1,
              "pattern": "[!-~]+"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created item and collectionRevision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkMutationResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ACCOUNT_DISABLED; authenticated mutations may also return CSRF_INVALID for a missing/invalid token or forbidden Origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND: missing Workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "IDEMPOTENCY_KEY_REUSED or REVISION_CONFLICT on collection/position exhaustion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "QUOTA_EXCEEDED: delete a Link before creating another",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ReorderLinksRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "collectionRevision": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 0
          },
          "ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "collectionRevision",
          "ids"
        ]
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "fieldErrors",
          "message",
          "requestId"
        ]
      },
      "LinkListResponse": {
        "type": "object",
        "description": "Entire filtered collection; never paginated. At most 500 items and 8 MiB compact UTF-8. collectionRevision describes the global owned collection.",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LinkResponse"
            },
            "maxItems": 500
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "nextCursor": {
            "type": "null",
            "description": "Always null; no cursor pagination"
          },
          "collectionRevision": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 0,
            "readOnly": true
          }
        },
        "required": [
          "collectionRevision",
          "items",
          "nextCursor",
          "total"
        ]
      },
      "LinkResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "revision": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 1,
            "readOnly": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "label": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 300
          },
          "url": {
            "type": "string",
            "format": "uri",
            "maxLength": 2000
          },
          "scope": {
            "type": "string",
            "enum": [
              "all",
              "unity",
              "server"
            ]
          },
          "position": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 0,
            "readOnly": true
          }
        },
        "required": [
          "createdAt",
          "description",
          "id",
          "label",
          "position",
          "revision",
          "scope",
          "updatedAt",
          "url"
        ]
      },
...
      "TaskResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "revision": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 1,
            "readOnly": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "projectName": {
            "type": "string",
            "readOnly": true
          },
          "scope": {
            "type": "string",
            "enum": [
              "unity",
              "server"
            ],
            "readOnly": true
          },
          "description": {
            "type": "string",
            "maxLength": 10000
          },
          "status": {
            "type": "string",
            "enum": [
              "todo",
              "doing",
              "done"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "normal",
              "high"
            ]
          },
          "tag": {
            "type": "string",
            "maxLength": 40
          },
          "deletedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": [
          "createdAt",
          "deletedAt",
          "description",
          "id",
          "priority",
          "projectId",
          "projectName",
          "revision",
          "scope",
          "status",
          "tag",
          "title",
          "updatedAt"
        ]
      },
      "RestoreTaskRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "revision": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 1
          }
        },
        "required": [
          "revision"
        ]
      },
...
      "CreateLinkRequest": {
        "type": "object",
        "additionalProperties": false,
        "description": "Only supplied fields change. Null, unknown and duplicate fields reject. URLs must have an http/https host and no credentials; strings use UTF-16 lengths.",
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1
          },
          "description": {
            "type": "string",
            "default": "",
            "maxLength": 300,
            "minLength": 0
          },
          "url": {
            "type": "string",
            "format": "uri",
            "maxLength": 2000,
            "minLength": 1
          },
          "scope": {
            "type": "string",
            "default": "all",
            "enum": [
              "all",
              "unity",
              "server"
            ]
          }
        },
        "required": [
          "label",
          "url"
        ]
      },
      "LinkMutationResponse": {
        "type": "object",
        "properties": {
          "item": {
            "$ref": "#/components/schemas/LinkResponse"
          },
          "collectionRevision": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 0,
            "readOnly": true
          }
        },
        "required": [
          "collectionRevision",
          "item"
        ]
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "name": "JSESSIONID",
        "in": "cookie"
      }
    }
  }
}